move internal stuff to internal in client api move projections to .Projections
