+++ src/EventStore.ClientAPI/ClientOperations/CreatePersistentSubscriptionOperation.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using EventStore.ClientAPI.Common.Utils;
 using EventStore.ClientAPI.Exceptions;
+using EventStore.ClientAPI.Internal;
 using EventStore.ClientAPI.Messages;
 using EventStore.ClientAPI.SystemData;
