+++ src/EventStore.ClientAPI/EventStore.ClientAPI.csproj
@@ -53,14 +53,14 @@
     <Compile Include="AllCheckpoint.cs" />
     <Compile Include="AllEventsSlice.cs" />
     <Compile Include="ConnectionString.cs" />
-    <Compile Include="IPEndPointExtensions.cs" />
+    <Compile Include="Internal\IPEndPointExtensions.cs" />
     <Compile Include="UserManagement\ResetPasswordDetails.cs" />
     <Compile Include="UserManagement\ChangePasswordDetails.cs" />
     <Compile Include="EventStorePersistentSubscription.cs" />
     <Compile Include="Common\Log\FileLogger.cs" />
     <Compile Include="Exceptions\UserCommandConflictException.cs" />
     <Compile Include="Exceptions\UserCommandFailedException.cs" />
-    <Compile Include="PersistentSubscriptionCreateResult.cs" />
+    <Compile Include="Internal\PersistentSubscriptionCreateResult.cs" />
     <Compile Include="PersistentSubscriptionSettings.cs" />
     <Compile Include="PersistentSubscriptionSettingsBuilder.cs" />
     <Compile Include="ClientOperations\CreatePersistentSubscriptionOperation.cs" />
@@ -70,16 +70,16 @@
     <Compile Include="ClientOperations\VolatileSubscriptionOperation.cs" />
     <Compile Include="PersistentEventStoreSubscription.cs" />
     <Compile Include="ClientOperations\ConnectToPersistentSubscriptionOperation.cs" />
-    <Compile Include="Core\GossipSeedEndPointDiscoverer.cs" />
-    <Compile Include="Core\IEventStoreConnectionLogicHandler.cs" />
-    <Compile Include="Core\NodeEndPoints.cs" />
+    <Compile Include="Internal\GossipSeedEndPointDiscoverer.cs" />
+    <Compile Include="Internal\IEventStoreConnectionLogicHandler.cs" />
+    <Compile Include="Internal\NodeEndPoints.cs" />
     <Compile Include="GossipSeed.cs" />
-    <Compile Include="PersistentSubscriptionCreateStatus.cs" />
-    <Compile Include="PersistentSubscriptionDeleteResult.cs" />
-    <Compile Include="PersistentSubscriptionDeleteStatus.cs" />
+    <Compile Include="Internal\PersistentSubscriptionCreateStatus.cs" />
+    <Compile Include="Internal\PersistentSubscriptionDeleteResult.cs" />
+    <Compile Include="Internal\PersistentSubscriptionDeleteStatus.cs" />
     <Compile Include="PersistentSubscriptionNakEventAction.cs" />
-    <Compile Include="PersistentSubscriptionUpdateResult.cs" />
-    <Compile Include="PersistentSubscriptionUpdateStatus.cs" />
+    <Compile Include="Internal\PersistentSubscriptionUpdateResult.cs" />
+    <Compile Include="Internal\PersistentSubscriptionUpdateStatus.cs" />
     <Compile Include="StreamCheckpoint.cs" />
     <Compile Include="ClientAuthenticationFailedEventArgs.cs" />
     <Compile Include="ClientErrorEventArgs.cs" />
@@ -120,37 +120,37 @@
     <Compile Include="UserManagement\UserUpdateInformation.cs" />
     <Compile Include="UserManagement\UsersClient.cs" />
     <Compile Include="UserManagement\UsersManager.cs" />
-    <Compile Include="VolatileEventStoreSubscription.cs" />
+    <Compile Include="Internal\VolatileEventStoreSubscription.cs" />
     <Compile Include="WriteResult.cs" />
     <Compile Include="DeleteResult.cs" />
     <Compile Include="SystemSettings.cs" />
     <Compile Include="EventReadResult.cs" />
-    <Compile Include="Core\EventStoreConnectionLogicHandler.cs" />
-    <Compile Include="Core\IEndPointDiscoverer.cs" />
-    <Compile Include="Core\OperationsManager.cs" />
-    <Compile Include="Core\StaticEndPointDiscoverer.cs" />
-    <Compile Include="Core\SubscriptionsManager.cs" />
-    <Compile Include="EventStoreNodeConnection.cs" />
+    <Compile Include="Internal\EventStoreConnectionLogicHandler.cs" />
+    <Compile Include="Internal\IEndPointDiscoverer.cs" />
+    <Compile Include="Internal\OperationsManager.cs" />
+    <Compile Include="Internal\StaticEndPointDiscoverer.cs" />
+    <Compile Include="Internal\SubscriptionsManager.cs" />
+    <Compile Include="Internal\EventStoreNodeConnection.cs" />
     <Compile Include="Exceptions\ClusterException.cs" />
     <Compile Include="Exceptions\NotAuthenticatedException.cs" />
     <Compile Include="IEventStoreConnection.cs" />
-    <Compile Include="Core\IEventStoreTransactionConnection.cs" />
+    <Compile Include="Internal\IEventStoreTransactionConnection.cs" />
     <Compile Include="StreamAcl.cs" />
     <Compile Include="SubscriptionDropReason.cs" />
     <Compile Include="Exceptions\AccessDeniedException.cs" />
     <Compile Include="StreamMetadata.cs" />
-    <Compile Include="Core\ClusterDnsEndPointDiscoverer.cs" />
-    <Compile Include="Core\Messages.cs" />
-    <Compile Include="Core\SimpleQueuedHandler.cs" />
-    <Compile Include="Consts.cs" />
+    <Compile Include="Internal\ClusterDnsEndPointDiscoverer.cs" />
+    <Compile Include="Internal\Messages.cs" />
+    <Compile Include="Internal\SimpleQueuedHandler.cs" />
+    <Compile Include="Internal\Consts.cs" />
     <Compile Include="EventStoreConnection.cs" />
     <Compile Include="EventStoreCatchUpSubscription.cs" />
-    <Compile Include="Empty.cs" />
+    <Compile Include="Internal\Empty.cs" />
     <Compile Include="EventData.cs" />
     <Compile Include="EventStoreSubscription.cs" />
     <Compile Include="Exceptions\EventStoreConnectionException.cs" />
     <Compile Include="Exceptions\ServerErrorException.cs" />
-    <Compile Include="ProjectionsClient.cs" />
+    <Compile Include="Projections\ProjectionsClient.cs" />
     <Compile Include="ReadDirection.cs" />
     <Compile Include="ResolvedEvent.cs" />
     <Compile Include="Exceptions\CannotEstablishConnectionException.cs" />
@@ -174,7 +174,7 @@
     <Compile Include="Messages\ClientMessagesExtensions.cs" />
     <Compile Include="Messages\ClusterMessages.cs" />
     <Compile Include="Position.cs" />
-    <Compile Include="ProjectionsManager.cs" />
+    <Compile Include="Projections\ProjectionsManager.cs" />
     <Compile Include="RecordedEvent.cs" />
     <Compile Include="SliceReadStatus.cs" />
     <Compile Include="StreamPosition.cs" />
