+++ src/EventStore.ClientAPI/Internal/Consts.cs
