+++ src/EventStore.ClientAPI/Internal/GossipSeedEndPointDiscoverer.cs
