+++ src/EventStore.ClientAPI/Internal/IEndPointDiscoverer.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net;
 using System.Threading.Tasks;
 
-namespace EventStore.ClientAPI.Core
+namespace EventStore.ClientAPI.Internal
 {
     internal interface IEndPointDiscoverer
     {
