+++ src/EventStore.ClientAPI/Internal/Messages.cs
@@ -6,7 +6,7 @@ using EventStore.ClientAPI.Common.Utils;
 using EventStore.ClientAPI.SystemData;
 using EventStore.ClientAPI.Transport.Tcp;
 
-namespace EventStore.ClientAPI.Core
+namespace EventStore.ClientAPI.Internal
 {
     internal abstract class Message
     {
