+++ src/EventStore.ClientAPI/Internal/PersistentSubscriptionCreateStatus.cs
@@ -1,9 +1,9 @@
-namespace EventStore.ClientAPI
+namespace EventStore.ClientAPI.Internal
 {
     /// <summary>
     /// Enumeration representing the status of a single subscription create message.
     /// </summary>
-    public enum PersistentSubscriptionUpdateStatus
+    enum PersistentSubscriptionCreateStatus
     {
         /// <summary>
         /// The subscription was created successfully
@@ -17,9 +17,5 @@ namespace EventStore.ClientAPI
         /// Some failure happened creating the subscription
         /// </summary>
         Failure = 2,
-        /// <summary>
-        /// You do not have permissions to update this subscription
-        /// </summary>
-        AccessDenied = 3,
     }
 }
