+++ src/EventStore.ClientAPI/Internal/PersistentSubscriptionDeleteResult.cs
@@ -1,4 +1,4 @@
-namespace EventStore.ClientAPI
+namespace EventStore.ClientAPI.Internal
 {
     /// <summary>
     /// A Persistent Subscription Delete Result is the result of a single operation deleting a
