+++ src/EventStore.ClientAPI/RecordedEvent.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Text;
+using EventStore.ClientAPI.Internal;
 using EventStore.ClientAPI.Messages;
 
 namespace EventStore.ClientAPI
