+++ src/EventStore.ClientAPI/StreamEventsSlice.cs
@@ -1,4 +1,5 @@
 using EventStore.ClientAPI.Common.Utils;
+using EventStore.ClientAPI.Internal;
 using EventStore.ClientAPI.Messages;
 
 namespace EventStore.ClientAPI
