+++ src/EventStore.Core.Tests/ClientAPI/soft_delete.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading;
 using EventStore.ClientAPI;
 using EventStore.ClientAPI.Exceptions;
+using EventStore.ClientAPI.Internal;
 using EventStore.Core.Tests.ClientAPI.Helpers;
 using EventStore.Core.Tests.Helpers;
 using NUnit.Framework;
